requirements.txt
fastapi
uvicorn
requests
faiss-cpu
python-docx
pdfminer.six
pillow
pytesseract
qdrant-client
dotenv
ollama
aiohttp
azure-storage-blob
jwt
PyJWT
cryptography